+++ plugins/job/src/main/kotlin/fr/pickaria/job/Main.kt
@@ -1,7 +1,6 @@
 package fr.pickaria.job
 
 import fr.pickaria.menu.menuController
-import org.bukkit.Material
 import org.bukkit.plugin.java.JavaPlugin
 
 lateinit var jobController: JobController
