+++ plugins/menu/src/main/kotlin/fr/pickaria/menu/MenuItem.kt
@@ -5,7 +5,6 @@ import org.bukkit.enchantments.Enchantment
 import org.bukkit.event.inventory.InventoryClickEvent
 import org.bukkit.inventory.ItemFlag
 import org.bukkit.inventory.ItemStack
-import java.util.function.Consumer
 
 class MenuItem {
 	var name: String = ""
